+++ installers/common/WindowsInstaller.hs
@@ -225,12 +225,12 @@ writeInstallerNSIS outName (Version fullVersion') InstallerConfig{installDirecto
                     file [] "topology.yaml"
                     file [] "genesis.json"
                     when (clusterName /= Selfnode) $ do
+                      file [] "genesis-conway.json"
                       file [] "genesis-byron.json"
                       file [] "genesis-shelley.json"
                       file [] "genesis-alonzo.json"
                     file [] "libsodium-23.dll"
-                    file [] "libsecp256k1-1.dll"
-                    file [] "libsecp256k1-0.dll"
+                    file [] "libsecp256k1-2.dll"
                     file [] "libssl-3-x64.dll"
                     file [] "libcrypto-3-x64.dll"
                     when (clusterName == Selfnode) $ do